+++ src/main/kotlin/com/jaoafa/vcspeaker/api/update/ReloadModule.kt
@@ -1,0 +1,346 @@
+package com.jaoafa.vcspeaker.api.update.modules
+
+import com.jaoafa.vcspeaker.KordStarter
+import com.jaoafa.vcspeaker.VCSpeaker
+import com.jaoafa.vcspeaker.api.update.UpdateServerType
+import com.jaoafa.vcspeaker.api.update.UpdateServerTypePlugin
+import com.jaoafa.vcspeaker.api.update.invalidTypeKey
+import com.jaoafa.vcspeaker.api.update.types.InitFinishedRequest
+import com.jaoafa.vcspeaker.api.update.types.UpdateError
+import com.jaoafa.vcspeaker.reload.Reload
+import com.jaoafa.vcspeaker.reload.UpdateRequest
+import com.jaoafa.vcspeaker.reload.state.State
+import com.jaoafa.vcspeaker.reload.state.StateManager
+import com.jaoafa.vcspeaker.tts.providers.ProviderContext
+import com.jaoafa.vcspeaker.tts.providers.soundmoji.SoundmojiContext
+import com.jaoafa.vcspeaker.tts.providers.voicetext.VoiceTextContext
+import io.github.oshai.kotlinlogging.KotlinLogging
+import io.ktor.client.HttpClient
+import io.ktor.client.call.body
+import io.ktor.client.engine.cio.CIO
+import io.ktor.client.request.HttpRequestBuilder
+import io.ktor.client.request.accept
+import io.ktor.client.request.basicAuth
+import io.ktor.client.request.post
+import io.ktor.client.request.setBody
+import io.ktor.http.ContentType
+import io.ktor.http.HttpStatusCode
+import io.ktor.http.contentType
+import io.ktor.serialization.kotlinx.json.json
+import io.ktor.server.application.Application
+import io.ktor.server.application.ServerReady
+import io.ktor.server.application.install
+import io.ktor.server.auth.UserIdPrincipal
+import io.ktor.server.auth.authenticate
+import io.ktor.server.auth.authentication
+import io.ktor.server.auth.basic
+import io.ktor.server.plugins.contentnegotiation.ContentNegotiation
+import io.ktor.server.request.receiveText
+import io.ktor.server.response.respond
+import io.ktor.server.routing.RoutingCall
+import io.ktor.server.routing.get
+import io.ktor.server.routing.post
+import io.ktor.server.routing.route
+import io.ktor.server.routing.routing
+import kotlinx.coroutines.runBlocking
+import kotlinx.io.IOException
+import kotlinx.serialization.KSerializer
+import kotlinx.serialization.builtins.serializer
+import kotlinx.serialization.json.Json
+import kotlinx.serialization.modules.SerializersModule
+import kotlinx.serialization.modules.polymorphic
+import kotlinx.serialization.modules.subclass
+import java.security.SecureRandom
+import kotlin.io.encoding.Base64
+import kotlin.io.encoding.ExperimentalEncodingApi
+import kotlin.system.exitProcess
+
+class ReloadModule(
+    val type: UpdateServerType,
+    var targetToken: String? = null,
+    var targetId: String? = null,
+    val sendBackIntSignal: Boolean
+) {
+    private val logger = KotlinLogging.logger { }
+
+    val reloaderJsonFormat = Json {
+        explicitNulls = false
+        serializersModule = SerializersModule {
+            polymorphic(ProviderContext::class) {
+                subclass(SoundmojiContext::class)
+                subclass(VoiceTextContext::class)
+            }
+        }
+    }
+
+    var sequence = 0
+        private set
+
+    val selfId = Reload.serverIds.random()
+
+    @OptIn(ExperimentalEncodingApi::class)
+    val selfToken = run {
+        val random = SecureRandom()
+        val bytes = ByteArray(32)
+        random.nextBytes(bytes)
+        Base64.encode(bytes)
+    }
+
+    var targetPort = 0
+
+    val client = HttpClient(CIO) {
+        install(io.ktor.client.plugins.contentnegotiation.ContentNegotiation) {
+            json(reloaderJsonFormat)
+        }
+    }
+
+    fun requesterConfig(body: Any? = null): HttpRequestBuilder.() -> Unit = {
+        contentType(ContentType.Application.Json)
+        accept(ContentType.Application.Json)
+
+        basicAuth(
+            username = selfId,
+            password = targetToken ?: ""
+        )
+
+        if (body != null) setBody(body)
+    }
+
+    /**
+     * VCSpeaker の内部 API にリクエストを送信します。
+     *
+     * @param T Request body の型
+     * @param R Response body の型
+     * @param route API のルート。頭の / は不要です。
+     * @param data 任意の request body
+     * @return Response body
+     */
+    suspend inline fun <reified T> requestUpdate(
+        route: String,
+        data: T,
+        serializer: KSerializer<T>
+    ) {
+        val url = "http://localhost:$targetPort/$route"
+
+        val response = client.post(
+            url,
+            requesterConfig(
+                reloaderJsonFormat.encodeToString(
+                    UpdateRequest.serializer(serializer),
+                    UpdateRequest(sequence, data)
+                )
+            )
+        )
+
+
+        if (response.status != HttpStatusCode.OK) {
+            val error = response.body<UpdateError>()
+            throw IOException("Failed to request update: ${response.status} ${error.message}")
+        }
+
+        return
+    }
+
+    /**
+     * ルートで受信した UpdateRequest をデコードします。
+     *
+     * @param serializer Body のシリアライザ
+     * @return UpdateRequest<T>
+     */
+    suspend fun <T> RoutingCall.receiveUpdateOf(serializer: KSerializer<T>): UpdateRequest<T> {
+        val response = reloaderJsonFormat
+            .decodeFromString(UpdateRequest.serializer(serializer), receiveText())
+        return response
+    }
+
+    /**
+     * リクエストに肯定応答を返し、シーケンス番号を更新します。
+     *
+     * @param s シーケンス番号
+     */
+    private suspend fun RoutingCall.ok(s: Int) {
+        respond(HttpStatusCode.OK)
+        logger.info { "[S$s] Done." }
+        sequence = s + 1
+    }
+
+    fun Application.module() {
+        monitor.subscribe(ServerReady) {
+            if (sendBackIntSignal) {
+                runBlocking {
+                    requestUpdate(
+                        "update/current/init-finished",
+                        InitFinishedRequest(
+                            selfId,
+                            selfToken
+                        ),
+                        InitFinishedRequest.serializer()
+                    )
+                }
+            }
+        }
+
+        install(ContentNegotiation) {
+            json(reloaderJsonFormat)
+        }
+
+        install(UpdateServerTypePlugin) {
+            type = this@ReloadModule.type
+        }
+
+        authentication {
+            basic("update-basic-auth") {
+                realm = "VCSpeaker Updater API"
+                validate { credentials ->
+                    val namePass = if (targetId == null) true else { // Accept username only once
+                        credentials.name == targetId.toString()
+                    }
+
+                    if (namePass && credentials.password == selfToken) {
+                        UserIdPrincipal(credentials.name)
+                    } else null
+                }
+            }
+        }
+
+        routing {
+            get("/version") {
+                call.respond(HttpStatusCode.OK, VCSpeaker.version)
+            }
+
+            route("/update") {
+                authenticate("update-basic-auth") {
+                    route("/current") {
+                        /**
+                         * S0 - Latest から Current へ、初期化が完了したことを通知します。
+                         * Body: UpdateRequest<InitFinishedRequest>
+                         */
+                        post("/init-finished") {
+                            if (call.attributes[invalidTypeKey]) return@post
+
+                            val (s, request) = call.receiveUpdateOf(InitFinishedRequest.serializer())
+
+                            if (s != 0) {
+                                call.respond(HttpStatusCode.BadRequest, UpdateError("Sequence mismatch."))
+                                return@post
+                            }
+
+                            targetId = request.id
+                            targetToken = request.token
+
+                            call.ok(s)
+
+                            logger.info { "[S$sequence] $targetId has finished init. Transferring the state..." }
+
+                            try {
+                                requestUpdate(
+                                    "update/latest/state",
+                                    State.generate(),
+                                    State.serializer()
+                                )
+                            } catch (e: Exception) {
+                                logger.error(e) { "[S$sequence] Failed to transfer the state. Aborting update." }
+                                return@post
+                            }
+                        }
+
+                        /**
+                         * S2 - Latest から Current へ、ステート同期とログイン準備が完了したことを通知します。
+                         * Current は S3 - Ack を送信後に code 0 で終了します。
+                         * Body: UpdateRequest<Boolean>
+                         */
+                        post("/ready") {
+                            if (call.attributes[invalidTypeKey]) return@post
+
+                            val (s, _) = call.receiveUpdateOf(Boolean.serializer())
+
+                            if (s != 2) {
+                                call.respond(HttpStatusCode.BadRequest, UpdateError("Sequence mismatch."))
+                                return@post
+                            }
+
+                            call.ok(s)
+
+                            logger.info { "[S$sequence] $targetId is ready. Exiting..." }
+
+                            try {
+                                requestUpdate(
+                                    "update/latest/ack",
+                                    true,
+                                    Boolean.serializer()
+                                )
+                            } catch (e: Exception) {
+                                logger.error(e) { "[S$sequence] Failed to send ack signal. Aborting update. Exit cancelled." }
+                                return@post
+                            }
+
+                            VCSpeaker.removeShutdownHook()
+                            exitProcess(0)
+                        }
+                    }
+
+                    route("/latest") {
+                        /**
+                         * S1 - Current から Latest へ、[State] を転送します。
+                         * Body: UpdateRequest<State>
+                         */
+                        post("/state") {
+                            if (call.attributes[invalidTypeKey]) return@post
+
+                            val (s, state) = call.receiveUpdateOf(State.serializer())
+
+                            if (s != 1) {
+                                call.respond(HttpStatusCode.BadRequest, UpdateError("Sequence mismatch."))
+                                return@post
+                            }
+
+                            logger.info { "[S$sequence] $targetId has frozen the state. Restoring..." }
+
+                            StateManager.restore(state)
+
+                            call.ok(s)
+
+                            try {
+                                requestUpdate(
+                                    "update/current/ready",
+                                    true,
+                                    Boolean.serializer()
+                                )
+                            } catch (e: Exception) {
+                                logger.error(e) { "[S$sequence] Failed to send ready signal. Aborting update." }
+                                return@post
+                            }
+                        }
+
+                        /**
+                         * S3 - Current から Latest へ、肯定応答を送信し、Latest はログインを開始します。
+                         * Body: UpdateRequest<State>
+                         */
+                        post("/ack") {
+                            if (call.attributes[invalidTypeKey]) return@post
+
+                            val (s, _) = call.receiveUpdateOf(Boolean.serializer())
+
+                            if (s != 3) {
+                                call.respond(HttpStatusCode.BadRequest, UpdateError("Sequence mismatch."))
+                                return@post
+                            }
+
+                            call.ok(s)
+
+                            logger.info { "Logging in..." }
+
+                            val instance = KordStarter.instance
+
+                            if (instance != null) {
+                                instance.start()
+                            } else {
+                                throw IllegalStateException("Unexpected error: KordEx instance is null.")
+                            }
+                        }
+                    }
+                }
+            }
+        }
+    }
+}
