+++ src/main/kotlin/com/jaoafa/vcspeaker/api/update/UpdateServerTypePlugin.kt
@@ -1,22 +1,22 @@
-package com.jaoafa.vcspeaker.api
+package com.jaoafa.vcspeaker.api.update
 
-import com.jaoafa.vcspeaker.api.types.UpdateError
+import com.jaoafa.vcspeaker.api.update.types.UpdateError
 import io.ktor.http.*
 import io.ktor.server.application.*
 import io.ktor.server.request.*
-import io.ktor.server.response.*
+import io.ktor.server.response.respond
 import io.ktor.util.*
 
-class ServerTypePluginConfig {
-    var type: ServerType = ServerType.Unknown
+class UpdateServerTypePluginConfig {
+    var type: UpdateServerType = UpdateServerType.Unknown
 }
 
 // If the wrong type of endpoint is called, this attribute will be set to true.
 val invalidTypeKey = AttributeKey<Boolean>("invalidTypeKey")
 
-val ServerTypePlugin = createApplicationPlugin(
+val UpdateServerTypePlugin = createApplicationPlugin(
     name = "ServerTypePlugin",
-    createConfiguration = ::ServerTypePluginConfig
+    createConfiguration = ::UpdateServerTypePluginConfig
 ) {
     onCall { call ->
         val path = call.request.uri
